@@ -1,25 +1,12 @@
-﻿<Project Sdk="Microsoft.NET.Sdk.Web">
+<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.0</TargetFramework>
-    <RootNamespace>TreeGridCore</RootNamespace>
+    <TargetFramework>net5.0</TargetFramework>
+    <RootNamespace>treegrid_core</RootNamespace>
   </PropertyGroup>
 
   <ItemGroup>
-    <None Remove="Report.txt" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <None Include="App_Data\*" CopyToPublishDirectory="PreserveNewest" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.AspNetCore.All" Version="2.0.9" />
-    <PackageReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Design" Version="2.0.4" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <DotNetCliToolReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Tools" Version="2.0.4" />
+    <Folder Include="App_Data\" />
   </ItemGroup>
 
 </Project>
