@@ -1,5 +1,5 @@
 ﻿@{
-    ViewBag.Title = "TreeGrid source";
+    ViewBag.Title = Model.ShortPath;
 }
 @model TreeGridCore.Models.FileXml
 
