@@ -11,7 +11,7 @@
 
 @RenderBody()
 
-<script src="~/lib/jquery/jquery.js" asp-append-version="true"></script>
+<script src="~/node_modules/jquery/dist/jquery.js" asp-append-version="true"></script>
 <script src="~/js/site.js" asp-append-version="true"></script>
 
 @RenderSection("Scripts", required: false)
